# Moved the OQ_SAMPLE_SOURCES functionality in the source_reader