This gives a 21x speedup in the slowest task for CHN when running the mosaic tests:
```
# before
| calc_2291, maxmem=11.3 GB  | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 2_489    | 902.0     | 721       |
| computing mean_std         | 767.5    | 0.0       | 45_821    |
| ClassicalCalculator.run    | 697.2    | 1_026     | 1         |

| operation-duration | counts | mean    | stddev | min     | max     | slowfac |
|--------------------+--------+---------+--------+---------+---------+---------|
| classical          | 721    | 3.45199 | 569%   | 0.32808 | 526.7   | 152.6   |

# after
| calc_2759, maxmem=8.6 GB   | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 1_924    | 31.8      | 718       |
| computing mean_std         | 735.0    | 0.0       | 44_865    |
| ClassicalCalculator.run    | 194.1    | 1_070     | 1         |

| operation-duration | counts | mean    | stddev | min     | max     | slowfac |
|--------------------+--------+---------+--------+---------+---------+---------|
| classical          | 718    | 2.67923 | 44%    | 0.10194 | 24.9    | 9.30638 |
```